+++ src/js/web/src/routes/_app/admin/teams/$teamId.tsx
@@ -530,6 +530,7 @@ function AdminTeamDetailPage() {
         teamId={teamId}
         currentName={data.name}
         currentDescription={data.description}
+        currentIsActive={data.isActive}
         open={editOpen}
         onOpenChange={setEditOpen}
       />